+++ docs/CHANGELOG.md
@@ -36,6 +36,7 @@ sections to include in release notes:
 - New `sipnet-debug-view` tool for visualizing SIPNET debug output files (#359)
 - Plant mortality check with event output on occurrence (#359)
 - Guard against negative mineral N due to volatilization and/or leaching (#375)
+- `lai` as an output column in `sipnet.out`
 
 ### Fixed